module7-task1: правит по ревью
